Moved status codes to a separate contract and changed functions signatures

+++ solidity/contracts/interfaces/ManageableMarketPlace.sol
@@ -2,11 +2,11 @@ pragma solidity ^0.5.8;
 
 
 interface ManageableMarketPlace {
-    function submitRequest(uint deadline) external returns (int status, uint requestID);
+    function submitRequest(uint deadline) external returns (uint8 status, uint requestID);
 
-    function closeRequest(uint requestIdentifier) external returns (int status);
+    function closeRequest(uint requestIdentifier) external returns (uint8 status);
 
-    function decideRequest(uint requestIdentifier, uint[] calldata /*acceptedOfferIDs*/) external returns (int status);
+    function decideRequest(uint requestIdentifier, uint[] calldata /*acceptedOfferIDs*/) external returns (uint8 status);
 
-    function deleteRequest(uint requestIdentifier) external returns (int status);
+    function deleteRequest(uint requestIdentifier) external returns (uint8 status);
 }
